UT for ThaliCordovaPlugin - Extracted mock classes
