+++ src/android/java/io/jxcore/node/ConnectionModel.java
@@ -4,6 +4,7 @@
 package io.jxcore.node;
 
 import android.util.Log;
+
 import java.util.HashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -259,20 +260,24 @@ public class ConnectionModel {
      * @return The socket thread or null if not found.
      */
     private synchronized SocketThreadBase findSocketThread(final String peerId, final boolean isIncoming) {
+        SocketThreadBase foundThread = null;
+
         if (isIncoming) {
             for (IncomingSocketThread incomingSocketThread : mIncomingSocketThreads) {
                 if (incomingSocketThread != null && incomingSocketThread.getPeerProperties().getId().equalsIgnoreCase(peerId)) {
-                    return incomingSocketThread;
+                    foundThread = incomingSocketThread;
+                    break;
                 }
             }
         } else {
             for (OutgoingSocketThread outgoingSocketThread : mOutgoingSocketThreads) {
                 if (outgoingSocketThread != null && outgoingSocketThread.getPeerProperties().getId().equalsIgnoreCase(peerId)) {
-                    return outgoingSocketThread;
+                    foundThread = outgoingSocketThread;
+                    break;
                 }
             }
         }
 
-        return null;
+        return foundThread;
     }
 }
