+++ src/android/test/io/jxcore/node/InputStreamMock.java
@@ -1,0 +1,12 @@
+package io.jxcore.node;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+public class InputStreamMock extends InputStream {
+
+    @Override
+    public int read() throws IOException {
+        return 0;
+    }
+}
