+++ src/android/test/io/jxcore/node/JXcoreThaliCallbackMock.java
@@ -1,0 +1,4 @@
+package io.jxcore.node;
+
+public class JXcoreThaliCallbackMock extends JXcoreThaliCallback{
+}
