+++ src/android/test/io/jxcore/node/ListenerMock.java
@@ -1,0 +1,21 @@
+package io.jxcore.node;
+
+import io.jxcore.node.SocketThreadBase;
+
+public class ListenerMock implements SocketThreadBase.Listener {
+
+    @Override
+    public void onListeningForIncomingConnections(int portNumber) {
+
+    }
+
+    @Override
+    public void onDataTransferred(int numberOfBytes) {
+
+    }
+
+    @Override
+    public void onDisconnected(SocketThreadBase who, String errorMessage) {
+
+    }
+}
